requirements.txt
absl-py==2.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
commonmark==0.9.1
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
IQA-pytorch==0.1
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
networkx==3.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pandas==1.3.5
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scipy==1.11.2
seaborn==0.12.0
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tifffile==2023.9.18
torch @ http://download.pytorch.org/whl/cu113/torch-1.12.1%2Bcu113-cp39-cp39-linux_x86_64.whl#sha256=01854806973840bdf3fc0efc84811539cb37101afd6498ef24a7d7f1af3cb010
torchvision @ https://download.pytorch.org/whl/cu113/torchvision-0.13.1%2Bcu113-cp39-cp39-linux_x86_64.whl#sha256=2fa135937c65b1bb8e3bf8f620c4fb755c0cd02b1f57ba1d6f9c09467a9b70cf
tqdm==4.66.1
typing_extensions==4.8.0
unfoldNd==0.2.0
urllib3==1.26.16
Werkzeug==2.3.7
zipp==3.17.0
